rotki-bin: init at 1.32.1
I had a quick go at building it from source following [the guide]
[1], then decided on just wrapping the appimage after running into
some python build issues related to the `ruff` linter that was being
invoked in a build phase for some reason. I'm sure it's doable with some
patching if someone's keen enough, but this at least makes the package
available for now.

[1]: https://rotki.readthedocs.io/en/latest/installation_guide.html#build-from-source