Make overlay definition fail gracefully for unknown systems.
nixpkgs-24.05, even when system = "x86_64-linux" will instantiate a
nixpkgs configuration for i686-linux. I believe that this is because
of the stdenv bootstrapping in <nixpkgs/pkgs/stdenv/linux/default.nix>
references an i686-linux version of glibc.

Because modules/testing.nix references the overlay definition, and
because of the bootstrapping logic above, we need to make the flake
overlay definition to handle unknown systems more gracefully. Return
an empty set, in case we don't know the system.