Move repo to nixpkgs-24.05 (#523)
* mev-boost-prysm: Remove because of upstream deprecation.

* Introduce nixpkgs-23.11 and pkgs2311.

When we bump nixpkgs to 24.05 some packages need to stay behind, as
they don't compile with 24.05.

* ssvnode: Pin to 23.11, as ssvnode will not compile with 24.05.

* sedge: Pin to nixpkgs-23.11 as it does not compile with nixpkgs-24.05

* rotki-bin: Pin to 23.11 as it does not compile with 24.05.

* eth-validator-watcher: Pin to 23.11, as it won't compile with 24.05.

* Make overlay definition fail gracefully for unknown systems.

nixpkgs-24.05, even when system = "x86_64-linux" will instantiate a
nixpkgs configuration for i686-linux. I believe that this is because
of the stdenv bootstrapping in <nixpkgs/pkgs/stdenv/linux/default.nix>
references an i686-linux version of glibc.

Because modules/testing.nix references the overlay definition, and
because of the bootstrapping logic above, we need to make the flake
overlay definition to handle unknown systems more gracefully. Return
an empty set, in case we don't know the system.

* Move to nixpkgs-24.05.

* flake.lock update

* Remove mdDoc treewide, as it is a deprecated option

* reth: Fix broken syntax.

* geth: update vendorHash